🎉 feat: initialize the project with basic settings
The Typescript transpiler was configured with the output directory as "./build" and using Commonjs for module resolution.

The Jest test module was configured to use the ts-jest adapter module in order to automatically perform the transpilation process for each test. The "Coverage" coverage test was also configured. The tests were stored following the community's standard folder structure.

For the quality of the written code, the Prettier and ESLINT modules were used. Following the main recommendations provided by the "npx eslint ---init" command and adapting to the project context.