✨ feat(api): add classes to communicate with the NVD external API
The communication functionality for the API was created based on the Axios module. Thus, the "axiosBaseApi" class is responsible for exporting and configuring the Axios instance. The "nvdApiClient" class provides communication services with the NVD API for endpoints related to the CPE identifier dictionary and for querying registered CVE vulnerabilities.