✨ feat(factory): add a class for vulnerability manufacturing
This class aims to store and manage a list of structured vulnerabilities.

The main method "Create" aims to create a new object with a defined type from a json data object. This method uses a data object found in the list of vulnerabilities contained in the “vulnerabilities” property of the JSON response of a query to the NVD CVE API.