✨ feat(entities): add IScenarios interface and Scenarios class
The interface describes the main public functionalities for the Scenarios class.

The Scenarios class aims to evaluate the presence of a vulnerability in a software. To do this, a structure is used that establishes different criteria for different environments to be considered vulnerable, in addition to characteristics of the execution environment, in order to find a scenario in which there is a match.

Main methods of the class:
- isCpeVulnerable;
- extractVersionProduct.