✨ feat(services/types): add Report and Vulnerability types
The files containing the types used to support the services have been added.
The Report file contains the types that describe the data resulting from the output of the assessment service and the structures that make up the final report.
The Vulnerability file contains the data structure that describes the final result of the vulnerability search for a CPE identifier.