✨ feat: add main process
This class contains the method responsible for executing the process of extracting and analyzing the information present in the RPKI environment and persisting the information in the Report object. The extraction process starts with the operating system information and if it is not possible to find it, the user will be contacted to provide such information or cancel the process. All collected information is persisted in the instance received as a parameter. The kitty methods for starting the process and the method for exporting the file in ".json" format were also implemented.