✨ feat: add start.py file
The file is intended to be a wrapper to initialize the entire software. It has a main function, which adjusts the interpreter path so that it correctly performs imports when the application is in the /opt/local/bin/mirak-extractor installation folder. Finally, the main function loads the module present in the cli.py file