✨ feat(src/shared): added functionality to extract the CPE version
Extract the CPE version so that it can be numerically comparable. To do this, it was established that each value between the points will be an item added to a return list, with any non-numeric value being ignored and any numeric value separated from the points being added to the last value in the list. Finally, if there is only the identifier of any “*”, a list containing the largest possible integer value is returned.