♻️ refactor(service): remove progress bar, now managed by another class
The progress bar was removed from this class, as it is now managed by the EvaluationService class, which is now responsible for this behavior.

This improves the organization of the code, separating responsibilities and facilitating maintenance. And it allows for better usability for the process.

No functionality was lost, only the display of the bar was relocated to the new class.